requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
attrs==25.3.0
bcrypt==3.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cloudinary==1.44.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
distlib==0.3.9
dlib==20.0.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.8
fastapi-cli==0.0.7
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.8
fpdf==1.7.2
frozenlist==1.5.0
fuzzywuzzy==0.18.0
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
joblib==1.4.2
josh==0.1.0
jsonpickle==4.0.5
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
Levenshtein==0.27.1
lxml==5.3.1
markdown-it-py==3.0.0
markovify==0.9.4
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.2.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
opencv-python==4.12.0.88
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.48
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.1
pyproject-toml==0.1.0
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-Levenshtein==0.27.1
python-louvain==0.16
python-multipart==0.0.20
pytz==2024.2
pyvis==0.3.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
qr==0.6.0
qrcode==8.0
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
selenium==4.30.0
setuptools==80.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.42
sqlmodel==0.0.24
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.3
textblob==0.19.0
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
twilio==9.7.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.4.0
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.3
watchfiles==1.0.4
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
wheel==0.45.1
wordcloud==1.9.4
wsproto==1.2.0
yarl==1.18.3
